Dive into Python 3: Chapter 2 and 3
Exercise

- Chapter 2: Native Datatypes
- Chapter 3: Comprehensions

Write a Python program to check whether a given number is a narcissistic
number or not. Narcissistic numbers are numbers that are the sum of
their digits when raised to the power of the number of digits. For
example, 371 is a narcissistic number; it has three digits, and if we
cube each digits 33 + 73 + 13 the sum is 371.